+++ GerenciamentoMecanicaSistema/Contracts/Requests/Catalog/CreateServiceRequest.cs
@@ -5,7 +5,7 @@ using System.ComponentModel.DataAnnotations;
 
 namespace GerenciamentoMecanicaSistema.Contracts.Requests.Catalog
 {
-    public class CreateServiceRequest(string description, float hours, double pricePerHour, int amount)
+    public class CreateServiceRequest(string description, float hours, decimal pricePerHour, int amount)
     {
         [Description("Descrição do serviço")]
         [Required, RegularNonEmptyStringExpression]
@@ -17,7 +17,7 @@ namespace GerenciamentoMecanicaSistema.Contracts.Requests.Catalog
 
         [Description("Preço por hora do serviço")]
         [Required, GenericValueValidation]
-        public double PricePerHour { get; set; } = pricePerHour;
+        public decimal PricePerHour { get; set; } = pricePerHour;
 
         [Description("Quantidade de serviços")]
         [Required, GenericValueValidation]
