+++ Service.Interface/Commands/Catalog/CreateServiceCommand.cs
@@ -1,4 +1,4 @@
 namespace Service.Interface.Commands.Catalog
 {
-    public record CreateServiceCommand(string Description, float Hours, double PricePerHour, int Amount);
+    public record CreateServiceCommand(string Description, float Hours, decimal PricePerHour, int Amount);
 }
