+++ Service.Interface/Commands/Stock/CreateMaterialCommand.cs
@@ -1,4 +1,4 @@
 namespace Service.Interface.Commands.Stock
 {
-    public record CreateMaterialCommand(string Name, string Brand, double Price, int Amount);
+    public record CreateMaterialCommand(string Name, string Brand, decimal Price, int Amount);
 }
