+++ Service.Interface/Results/Catalog/ServiceResult.cs
@@ -1,4 +1,4 @@
 namespace Service.Interface.Results.Catalog
 {
-    public record ServiceResult(Guid Id, string Description, float Hours, double PricePerHour, int Amount);
+    public record ServiceResult(Guid Id, string Description, float Hours, decimal PricePerHour, int Amount);
 }
