+++ Service.Interface/Results/Order/DetailedWorkOrderResult.cs
@@ -4,7 +4,7 @@ using Service.Interface.Results.Stock;
 
 namespace Service.Interface.Results.Order
 {
-    public record DetailedWorkOrderResult(Guid Id, string CustomerDocument, string VehicleLicensePlate, double Budget, string Status, DateTime DateCreated, DateTime DateFinished, List<ServiceResult> Services, List<MaterialResult> Materials, TimeSpan Duration)
+    public record DetailedWorkOrderResult(Guid Id, string CustomerDocument, string VehicleLicensePlate, decimal Budget, string Status, DateTime DateCreated, DateTime DateFinished, List<ServiceResult> Services, List<MaterialResult> Materials, TimeSpan Duration)
     {
         public static DetailedWorkOrderResult Create(IOrder order)
         {
